Handle overflows when getting the idle duration
The GetTickCount method produces an overflow after around 40
days. This can lead to some interesting side effects. Since we only
used this value as a substitute for the datetime.now() value, we can
substitute the call by storing the datetime of when we noticed a
change in the last_input_info tick.